Made delete response same as node.

+++ todo/src/main/java/com/statsig/todo/services/TodoService.java
@@ -1,14 +1,13 @@
 package com.statsig.todo.services;
 
 
+import com.google.gson.JsonObject;
 import com.statsig.todo.model.Todo;
 import com.statsig.todo.repository.TodoRepository;
 import org.springframework.stereotype.Service;
 
 import java.time.ZonedDateTime;
-import java.util.Date;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 @Service
 public class TodoService {
@@ -32,8 +31,16 @@ public class TodoService {
         return todoRepository.save(todo);
     }
 
-    public void deleteTodo(Long id) {
-        todoRepository.deleteById(id);
+    public HashMap deleteTodo(Long id) {
+        HashMap<String,String> map = new HashMap<>();
+        Todo todo = getTodoById(id);
+        if(!Objects.isNull(todo)){
+            map.put("id",String.valueOf(todo.getId()));
+            todoRepository.deleteById(id);
+        }else{
+            map.put("error","Failed to delete TODO");
+        }
+        return map;
     }
 
     public Todo updateTodo(Todo todo) {
